Move all the `withSubBlock` HOC into the section block edit
